--- a/src/hotspot/cpu/aarch64/matcher_aarch64.hpp
+++ b/src/hotspot/cpu/aarch64/matcher_aarch64.hpp
@@ -139,8 +139,9 @@
   }
 
   // Does the CPU supports vector unsigned comparison instructions?
-  static constexpr bool supports_vector_comparison_unsigned(int vlen, BasicType bt) {
-    return false;
+  static const bool supports_vector_comparison_unsigned(int vlen, BasicType bt) {
+    // Not supported on SVE yet.
+    return !UseSVE;
   }
 
   // Some microarchitectures have mask registers used on vectors