… test/hotspot/gtest/aarch64/aarch64-asmtest.py
… test/hotspot/gtest/aarch64/asmtest.out.h