This patch implements unsigned vector comparison on AArch64. The
performance of unsigned comparison improves about 4x~5x in my local with
Byte128Vector.java[1].

Before:
Benchmark                               Score(op/ms)     Error
Byte128Vector.UNSIGNED_GE#size(1024)    99.953           6.17
Byte128Vector.UNSIGNED_GT#size(1024)    95.334           8.865
Byte128Vector.UNSIGNED_LE#size(1024)    76.908           24.332
Byte128Vector.UNSIGNED_LT#size(1024)    78.362           23.507

After:
Benchmark                               Score(op/ms)     Error
Byte128Vector.UNSIGNED_GE#size(1024)    421.809          25.57
Byte128Vector.UNSIGNED_GT#size(1024)    420.653          26.779
Byte128Vector.UNSIGNED_LE#size(1024)    316.754          92.889
Byte128Vector.UNSIGNED_LT#size(1024)    423.683          26.508

[Test]
- All vector API test cases passed without new failure. 8265312[2] has
  been implemented this on x86 and supplied sufficient test cases for
  all kinds of vector.
- No performance regression for other comparisons.
- libjvm.so drops off about 200KB after this patch by combining those
  vector compare rules.

[1] https://github.com/openjdk/panama-vector/blob/vectorIntrinsics/test/jdk/jdk/incubator/vector/benchmark/src/main/java/benchmark/jdk/incubator/vector/Byte128Vector.java#L1198
[2] https://github.com/openjdk/panama-vector/pull/68

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8267663](https://bugs.openjdk.java.net/browse/JDK-8267663): [vector] Add unsigned comparison operators on AArch64


### Reviewers
 * [Andrew Haley](https://openjdk.java.net/census#aph) (@theRealAph - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/4358/head:pull/4358` \
`$ git checkout pull/4358`

Update a local copy of the PR: \
`$ git checkout pull/4358` \
`$ git pull https://git.openjdk.java.net/jdk pull/4358/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 4358`

View PR using the GUI difftool: \
`$ git pr show -t 4358`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/4358.diff">https://git.openjdk.java.net/jdk/pull/4358.diff</a>

</details>